5/5.2/requirements.txt
psycopg2-binary==2.9.7
kafka-python==2.0.2
clickhouse-driver==0.2.6
python-dotenv==1.0.0 